+++ src/Relude/Extra/Foldable1.hs
@@ -30,6 +30,8 @@ class Foldable f => Foldable1 f where
 
     >>> foldMap1 SG.Sum (1 :| [2, 3, 4])
     Sum {getSum = 10}
+    >>> foldMap1 show (123 :| [456, 789, 0])
+    "1234567890"
     -}
     foldMap1 :: Semigroup m => (a -> m) -> f a -> m
 
@@ -88,9 +90,11 @@ instance Foldable1 NonEmpty where
     fold1 = sconcat
     {-# INLINE fold1 #-}
 
-    foldMap1 :: Semigroup m => (a -> m) -> NonEmpty a -> m
-    foldMap1 f (a :| [])     = f a
-    foldMap1 f (a :| b : bs) = f a <> foldMap1 f (b :| bs)
+    foldMap1 :: forall m a . Semigroup m => (a -> m) -> NonEmpty a -> m
+    foldMap1 f (a :| as) = foldr go f as a
+      where
+        go :: a -> (a -> m) -> a -> m
+        go b g x = f x <> g b
     {-# INLINE foldMap1 #-}
 
     toNonEmpty :: NonEmpty a -> NonEmpty a